+++ Vostok.Metrics.System/Host/HostMetricsCollector.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using JetBrains.Annotations;
+using Vostok.Metrics.System.Helpers;
 
 namespace Vostok.Metrics.System.Host
 {
@@ -9,23 +10,46 @@ namespace Vostok.Metrics.System.Host
     /// <para>It is designed to be invoked periodically.</para>
     /// </summary>
     [PublicAPI]
-    public class HostMetricsCollector
+    public class HostMetricsCollector : IDisposable
     {
+        private static readonly TimeSpan CacheTTL = TimeSpan.FromMilliseconds(100);
+
         private readonly Action<HostMetrics> nativeCollector;
+        private readonly Action disposeNativeCollector;
         private readonly DiskSpaceCollector diskSpaceCollector = new DiskSpaceCollector();
         private readonly TcpStateCollector tcpStateCollector = new TcpStateCollector();
+        private readonly ThrottlingCache<HostMetrics> cache;
+
+        public void Dispose()
+        {
+            disposeNativeCollector?.Invoke();
+            diskSpaceCollector?.Dispose();
+        }
 
         public HostMetricsCollector()
         {
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-                nativeCollector = new NativeHostMetricsCollector_Windows().Collect;
+            {
+                var collector = new NativeHostMetricsCollector_Windows();
+                nativeCollector = collector.Collect;
+                disposeNativeCollector = collector.Dispose;
+            }
 
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-                nativeCollector = new NativeHostMetricsCollector_Linux().Collect;
+            {
+                var collector = new NativeHostMetricsCollector_Linux();
+                nativeCollector = collector.Collect;
+                disposeNativeCollector = collector.Dispose;
+            }
+
+            cache = new ThrottlingCache<HostMetrics>(CollectInternal, CacheTTL);
         }
 
         [NotNull]
         public HostMetrics Collect()
+            => cache.Obtain();
+
+        private HostMetrics CollectInternal()
         {
             var metrics = new HostMetrics();
             CollectNativeMetrics(metrics);